chore: files renaming & relocation
