+++ frontend/src/components/datagrid/index.tsx
@@ -11,17 +11,10 @@ export * from './datagridTypes';
 
 import IdCell from './IdCell';
 import SortableHeaderRenderer from './SortableHeader';
-import { DefaultCellRenderers, createDefaultCellRenderers } from './DefaultCells';
-import DataGrid, { type DataGridProps } from './RawDataGrid';
+import { createDefaultCellRenderer } from './DefaultCellRenderer';
+import DataGrid, { type DataGridProps } from './DataGrid';
 
-export {
-  IdCell,
-  SortableHeaderRenderer,
-  DataGrid,
-  type DataGridProps,
-  DefaultCellRenderers,
-  createDefaultCellRenderers,
-};
+export { IdCell, SortableHeaderRenderer, DataGrid, type DataGridProps, createDefaultCellRenderer };
 
 // Export cell editors
 export * from './cell-editors';
