--- frontend/src/components/mcp/index.ts
@@ -1,4 +1,0 @@
-export { McpInstallation } from './McpInstallation';
-export { CursorDeeplinkGenerator } from './CursorDeeplinkGenerator';
-export type { MCPAgent, PlatformType } from '@/components/mcp/mcp-helper';
-export { MCP_AGENTS, createMCPConfig, createMCPServerConfig } from '@/components/mcp/mcp-helper';
