+++ frontend/src/features/onboard/components/mcp/index.ts
@@ -1,0 +1,4 @@
+export { McpInstallation } from './McpInstallation';
+export { CursorDeeplinkGenerator } from './CursorDeeplinkGenerator';
+export type { MCPAgent, PlatformType } from './mcp-helper';
+export { MCP_AGENTS, createMCPConfig, createMCPServerConfig } from './mcp-helper';
