+++ frontend/src/features/onboard/components/mcp/mcp-helper.ts
